pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ngwidgets"
keywords = ["nicegui", "components", "ui"]
description = "NiceGUI widgets and solution bazaar"
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"}
]
maintainers = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"},
]
readme = "README.md"
license= "Apache-2.0"
dependencies = [
	# https://nicegui.io/
    "nicegui>=2.10.1",
 	# orjson
 	# https://pypi.org/project/orjson/
 	"orjson>=3.9.10",
 	# https://pypi.org/project/beautifulsoup4/
 	"beautifulsoup4",
    # https://pypi.org/project/bcrypt/
    "bcrypt>=4.0.1",
	# https://pypi.org/project/tqdm/
	# progress bar
	"tdqm",
    # for debugging
	# https://pypi.org/project/pydevd/
	"pydevd>=2.10.0",
	# dateutil
	# https://pypi.org/project/python-dateutil/
	"python-dateutil>=2.8.2",
	# https://github.com/PyGithub/PyGithub
	# https://pypi.org/project/PyGithub/
	"PyGithub>=2.1.1",
	# https://pypi.org/project/pytz/
	"pytz",
	# https://pypi.org/project/dacite/
	"dacite>=1.8.1",
	# https://pypi.org/project/dataclasses-json/
    "dataclasses-json>=0.6.3",
    #https://pypi.org/project/PyYAML/
    "PyYAML>=6.0.1",
    #https://pypi.org/project/openai/
    "openai>=1.12.0",
    # https://pypi.org/project/colour/
	"colour>=0.1.5",
	# https://pypi.org/project/shutup/
	"shutup>=0.2.0",
	# gpxpy
	# https://github.com/tkrajina/gpxpy
	"gpxpy>=1.6.2"
]

requires-python = ">=3.9"
classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: Apache Software License"
]
dynamic = ["version"]

[tool.hatch.version]
path = "ngwidgets/__init__.py"

[project.urls]
Home = "https://github.com/WolfgangFahl/nicegui_widgets"
Documentation = "https://wiki.bitplan.com/index.php/nicegui_widgets"
Source = "https://github.com/WolfgangFahl/nicegui_widgets"

[project.optional-dependencies]
test = [
    "green",
    "selenium>=4.1.0",
    "webdriver-manager>=3.8.0",
    "pytest",
    "pytest-asyncio"
]
mbus = [
  # https://pypi.org/project/pyMeterBus/
  'pyMeterBus>=0.8.4',
]


[tool.hatch.build.targets.wheel]
only-include = ["ngwidgets"]

[tool.hatch.build.targets.wheel.sources]
"ngwidgets" = "ngwidgets"

[project.scripts]
ngdemo = "ngwidgets.ngwidgets_cmd:main"
niceuser = "ngwidgets.users:main"
aichat = "ngwidgets.ai_chat:run_chat"
gpxviewer = "ngwidgets.gpxviewer_server:main"
mbusviewer ="ngwidgets.mbus_viewer_cmd:main"
